# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PU Transit API
servers:
  - url: /api
    description: Base API path
security:
  - FirebaseBearer: []
...
paths:
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List routes
      responses:
        "200":
          description: Routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: saveRoute
      tags: [routes]
      summary: Create or update a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /routes/{id}:
    put:
      operationId: putRoute
      tags: [routes]
      summary: Create or replace a route under a client-chosen UUID (RTE-04) — idempotent, so a save whose response was lost can be retried without creating a second route; non-UUID ids are 400
      parameters:
        - $ref: "#/components/parameters/RouteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Replaced route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update a route by ID
      parameters:
        - $ref: "#/components/parameters/RouteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Permanently delete a never-published draft (RTE-03); published routes are archived instead (409 ROUTE_PUBLISHED)
      parameters:
        - $ref: "#/components/parameters/RouteId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /tracking:
    get:
      operationId: listTrackingFeeds
      tags: [tracking]
      summary: List current tracking feeds for administrators
      responses:
        "200":
          description: Tracking feeds keyed by bus ID
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/LiveTrackingFeed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}:
    get:
      operationId: getTrackingFeed
      tags: [tracking]
      summary: Get a sanitized tracking feed
      parameters:
        - $ref: "#/components/parameters/BusId"
      responses:
        "200":
          description: Sanitized tracking feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveTrackingFeed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/start:
    post:
      operationId: startTracking
      tags: [tracking]
      summary: Start or retry a tracking trip
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripStartInput"
      responses:
        "200":
          description: Idempotent active retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSession"
        "201":
          description: Started tracking trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/sample:
    post:
      operationId: submitTrackingSample
      tags: [tracking]
      summary: Publish a GPS sample for the current trip
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripSampleInput"
      responses:
        "200":
          description: Updated tracking trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/heartbeat:
    post:
      operationId: sendTrackingHeartbeat
      tags: [tracking]
      summary: Refresh the current tracking lease
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripHeartbeatInput"
      responses:
        "200":
          description: Updated tracking trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/end:
    post:
      operationId: endTracking
      tags: [tracking]
      summary: End or cancel a tracking trip
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripEndInput"
      responses:
        "200":
          description: End result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripEndResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/force-end:
    post:
      operationId: forceEndTracking
      tags: [tracking]
      summary: Force-end the current trip as an administrator
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForceEndRequest"
      responses:
        "200":
          description: Force-end result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForceEndResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/ServiceUnavailable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tracking/{busId}/handover:
    post:
      operationId: handoverTracking
      tags: [tracking]
      summary: Hand the bus to another driver in one confirmed action (ADM-10)
      description: Force-ends the current publisher and pre-authorizes the named driver's Start for today (IST) — through their standing bus, or a dated assignment on the given route — and writes one audit record naming both drivers. The previous phone drops to idle on its next upload (409 SESSION_CONFLICT).
      parameters:
        - $ref: "#/components/parameters/BusId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoverRequest"
      responses:
        "200":
          description: Handover result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoverResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/ServiceUnavailable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RouteId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    BusId:
      name: busId
      in: path
      required: true
      description: Canonical bus key (registration uppercased, letters and digits only)
      schema:
        type: string
        pattern: "^[A-Z0-9]{2,20}$"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Tracking ownership or replay conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Firebase JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Firebase email, membership, role, or assignment is not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Firebase or migration service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Route:
      type: object
      additionalProperties: false
      required: [id, shift, busNumber, origin, destination, stops, status]
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/RouteStatus"
        publishedVersion:
          description: Number of the version riders currently see (RTE-02); absent until the route is first published under versioning
          type: integer
          minimum: 1
        hasDraft:
          description: Admin lists only — the returned content is an unpublished draft; riders still see publishedVersion
          type: boolean
        pathSource:
          $ref: "#/components/schemas/RoutePathSource"
        shift:
          type: string
        busNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        stops:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/RouteStop"
        pathData:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/Coordinate"
        kind:
          $ref: "#/components/schemas/RouteKind"
    RouteKind:
      description: Regular city bus or a campus shuttle; absent means bus
      type: string
      enum: [bus, shuttle]
...
    RouteInput:
      type: object
      additionalProperties: false
      required: [shift, busNumber, origin, destination, stops, status]
      properties:
        id:
          type: string
          maxLength: 128
        status:
          description: The action - draft or published. Archiving is its own endpoint (RTE-03).
          type: string
          enum: [draft, published]
        pathSource:
          $ref: "#/components/schemas/RoutePathSource"
        shift:
          type: string
          minLength: 1
          maxLength: 200
        busNumber:
          type: string
          minLength: 1
          maxLength: 200
        origin:
          type: string
          minLength: 1
          maxLength: 200
        destination:
          type: string
          minLength: 1
          maxLength: 200
        stops:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/RouteStop"
        pathData:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/Coordinate"
        kind:
          $ref: "#/components/schemas/RouteKind"
...
    LiveTrackingFeed:
      type: object
      additionalProperties: false
      required:
        - protocolVersion
        - busId
        - tripId
        - generation
        - phase
        - requestedAt
        - startedAt
        - endedAt
        - heartbeatAt
        - gpsQuality
        - lastReportCapturedAt
        - lastReportReceivedAt
        - reportedAccuracy
        - lastValidCapturedAt
        - lastValidReceivedAt
        - location
        - status
        - serverTime
        - freshUntil
        - offlineAfter
        - direction
        - full
      properties:
        protocolVersion:
          type: integer
          enum: [2]
        direction:
          oneOf:
            - $ref: "#/components/schemas/TripDirection"
            - type: "null"
        full:
          description: Driver reports the bus has no free seats; false when the trip is not active
          type: boolean
        routeVersions:
          description: Route version pinned at Start for each published route of this bus (RTE-02); riders keep it until the trip ends. Empty for legacy routes and idle buses.
          type: object
          additionalProperties:
            type: integer
            minimum: 1
        busId:
          type: string
        tripId:
          type: ["string", "null"]
          format: uuid
        generation:
          type: integer
          minimum: 0
        phase:
          type: string
          enum: [not_started, active, ended]
        requestedAt:
          type: integer
          format: int64
          minimum: 0
        startedAt:
          type: integer
          format: int64
          minimum: 0
        endedAt:
          type: integer
          format: int64
          minimum: 0
        heartbeatAt:
          type: integer
          format: int64
          minimum: 0
        gpsQuality:
          type: string
          enum: [acquiring, good, weak, unavailable]
        lastReportCapturedAt:
          type: integer
          format: int64
          minimum: 0
        lastReportReceivedAt:
          type: integer
          format: int64
          minimum: 0
        reportedAccuracy:
          type: number
          minimum: 0
        lastValidCapturedAt:
          type: integer
          format: int64
          minimum: 0
        lastValidReceivedAt:
          type: integer
          format: int64
          minimum: 0
        location:
          type: ["object", "null"]
          additionalProperties: false
          required: [lat, lng, accuracy]
          properties:
            lat:
              type: number
              minimum: -90
              maximum: 90
            lng:
              type: number
              minimum: -180
              maximum: 180
            accuracy:
              type: number
              minimum: 0
              maximum: 100
        status:
          type: string
          enum: [not_started, acquiring, live, delayed, weak_gps, gps_unavailable, offline, ended]
        serverTime:
          type: integer
          format: int64
          minimum: 0
        freshUntil:
          type: integer
          format: int64
          minimum: 0
        offlineAfter:
          type: integer
          format: int64
          minimum: 0
    TripStartInput:
      type: object
      additionalProperties: false
      required: [tripId, publisherId, expectedGeneration, requestedAt]
      properties:
        tripId:
          type: string
          format: uuid
        publisherId:
          type: string
          format: uuid
        expectedGeneration:
          type: integer
          minimum: 0
        requestedAt:
          type: integer
          format: int64
          minimum: 0
        direction:
          $ref: "#/components/schemas/TripDirection"
...
    TripHeartbeatInput:
      type: object
      additionalProperties: false
      required: [tripId, publisherId, generation, sequence]
      properties:
        tripId:
          type: string
          format: uuid
        publisherId:
          type: string
          format: uuid
        generation:
          type: integer
          minimum: 1
        sequence:
          type: integer
          minimum: 0
        gpsUnavailable:
          type: boolean
        full:
          type: boolean
...
    TripSession:
      type: object
      additionalProperties: false
      required: [tripId, publisherId, generation, sequence, feed]
      properties:
        tripId:
          type: string
          format: uuid
        publisherId:
          type: string
          format: uuid
        generation:
          type: integer
          minimum: 1
        sequence:
          type: integer
          minimum: 0
        feed:
          $ref: "#/components/schemas/LiveTrackingFeed"
...
    ForceEndRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 200
...